The Terminal tab in Octorprint eroneously filters valid GCODE. This fixes that. Examples of code no currently accepted include:     + G38.2 ...     + M117testing [no space]     + G1X10 [no space]
Please note that, while it seems every 3D printer firmware around these
days is adopting the errant idea of using a single space to delimit GCODE
words, this is NOT copliant with GCODE standard RS274/NGC or any other gcode
standard known to me, for that matter. Whitespace -- all whitgespace, is to be
accepted and is ignored.  Even in the middle of a number.

In any case, imo, a GCODE terminal shouldn't be filtering ANYTHING. Such
filtering is not its domain, but rather that of the GCODE interpreter.
In fact, if we truly filtered for strictly only valid GCODE, then M117 would
immediately break -- and no doubt more. That command is not GCODE syntax
complaint. My argument here is, "If we're letting non-compliant code through,
then why do we attempt to filter anything at all, in the first palce?"

Well, I could go on and on and on some more. I'm an expert at it. But I think
I've said enough for now. :p

Cheers. :-)